+++ src/components/WizardSteps/interfaces.ts
@@ -1,0 +1,10 @@
+import { IWizardStepColorProps } from './WizardStep/interfaces';
+
+export interface IWizardStepProps {
+  active: boolean;
+  text: string;
+}
+
+export interface IWizardStepsProps extends IWizardStepColorProps {
+  wizardSteps: IWizardStepProps[];
+}