+++ src/components/WizardSteps/styles.ts
@@ -4,4 +4,8 @@ export const Container = styled.div`
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
+  justify-content: center;
+  padding-bottom: 2rem;
+  padding-top: 1rem;
+  width: 100%;
 `;