+++ CHANGELOG.md
@@ -19,6 +19,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Dropdown positions
 - Fixed Modal props' spreading
 - Fixed Modal overlay opacity
+## [1.22.0] - 2021-07-22
+
+- Renamed existing Breadcrumbs to WizardSteps
+- Fixed Public Profile Icon
+- Added new Breadcrumbs component
+
+## [1.21.3] - 2021-07-19
+
+- Fix ProductScan icon (again)
+
+## [1.21.2] - 2021-07-19
+
+- Fix ProductScan icon
 
 ## [1.21.1] - 2021-07-19